reorganizing tests structure
