+++ blog/tests/testPostForm.py
@@ -1,11 +1,10 @@
 from django.test import TestCase
-import unittest
 from django.contrib.auth.models import User
-from .models import Category, Post
-from .forms import LoginForm, PostForm
+from ..models import Category, Post
+from ..forms import LoginForm, PostForm
 from django.urls import reverse
     
-class PostFormTest(unittest.TestCase):
+class PostFormTest(TestCase):
     def setUp(self):
         self.user = User.objects.create_user(username="testuser", password="password")
         self.category = Category.objects.create(name="Technology")
@@ -68,6 +67,3 @@ class PostFormTest(unittest.TestCase):
     def test_view_uses_correct_template_register(self):
         response = self.client.get(reverse('register'))
         self.assertTemplateUsed(response, 'registration/register.html')
-
-if __name__ == '__main__':
-    unittest.main()
